requirements.txt
Django==5.2.4
gunicorn==22.0.0
psycopg2-binary==2.9.10
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
python-dotenv==1.1.1
django-environ==0.12.0
whitenoise==6.9.0
Pillow==11.3.0
python-slugify==8.0.4